requirements.txt
streamlit>=1.35.0
Pillow>=10.0.0
numpy>=1.24.0
transformers>=4.35.0
torch>=2.0.0
